requirements.txt
git+https://github.com/dolfies/discord.py-self
python-dotenv
google-generativeai
pillow
